Added file infrastructure for BMInterfacePlayer, and started shifting methods from BMInterface into BMInterfacePlayer
